MNT: Consolidate Calibrator code into detector folder
