--- data/calibrator_config.json
@@ -1,4 +1,0 @@
-{
-  "serial_identifier": "example_calibrator",
-  "data_path": "./example_data_path/"
-}
