--- data/config.json
@@ -1,8 +1,0 @@
-{
-  "serial_identifier": "example_detector",
-  "camera_implementation": "picamera",
-  "camera_connection": {
-    "usb_id": 1
-  },
-  "camera_intrinsic_parameters": {}
-}
