--- src/calibrator/__init__.py
@@ -1,1 +1,0 @@
-from .calibrator import Calibrator
