--- src/calibrator/fileio/__init__.py
@@ -1,1 +1,0 @@
-from .calibrator_configuration import CalibratorConfiguration
