+++ src/detector/__init__.py
@@ -1,2 +1,6 @@
+from .calibrator import Calibrator
 from .detector import Detector
-from .fileio import DetectorConfiguration
+from .interfaces import \
+    AbstractCameraInterface, \
+    AbstractMarkerInterface
+from .structures import DetectorConfiguration
