+++ src/detector/api/__init__.py
@@ -12,3 +12,21 @@ from .set_detection_parameters_request import SetDetectionParametersRequest
 from .set_marker_dictionary_request import SetMarkerDictionaryRequest
 from .start_capture_request import StartCaptureRequest
 from .stop_capture_request import StopCaptureRequest
+
+from .add_calibration_image_request import AddCalibrationImageRequest
+from .add_calibration_image_response import AddCalibrationImageResponse
+from .calibrate_request import CalibrateRequest
+from .calibrate_response import CalibrateResponse
+from .delete_staged_request import DeleteStagedRequest
+from .get_calibration_image_request import GetCalibrationImageRequest
+from .get_calibration_image_response import GetCalibrationImageResponse
+from .get_calibration_result_request import GetCalibrationResultRequest
+from .get_calibration_result_response import GetCalibrationResultResponse
+from .list_calibration_detector_resolutions_request import ListCalibrationDetectorResolutionsRequest
+from .list_calibration_detector_resolutions_response import ListCalibrationDetectorResolutionsResponse
+from .list_calibration_image_metadata_request import ListCalibrationImageMetadataRequest
+from .list_calibration_image_metadata_response import ListCalibrationImageMetadataResponse
+from .list_calibration_result_metadata_request import ListCalibrationResultMetadataRequest
+from .list_calibration_result_metadata_response import ListCalibrationResultMetadataResponse
+from .update_calibration_image_metadata_request import UpdateCalibrationImageMetadataRequest
+from .update_calibration_result_metadata_request import UpdateCalibrationResultMetadataRequest
