+++ src/detector/api/add_calibration_image_request.py
