+++ src/detector/api/add_calibration_image_response.py
