+++ src/detector/api/delete_staged_request.py
