+++ src/detector/api/get_calibration_image_request.py
