+++ src/detector/api/get_calibration_image_response.py
