+++ src/detector/api/list_calibration_detector_resolutions_response.py
