+++ src/detector/api/list_calibration_image_metadata_request.py
