+++ src/detector/api/list_calibration_image_metadata_response.py
