+++ src/detector/api/update_calibration_image_metadata_request.py
