+++ src/detector/api/update_calibration_result_metadata_request.py
