--- src/detector/fileio/__init__.py
@@ -1,2 +1,0 @@
-from .camera_connection import CameraConnection
-from .detector_configuration import DetectorConfiguration
