+++ src/detector/interfaces/__init__.py
@@ -1,0 +1,2 @@
+from .abstract_camera_interface import AbstractCameraInterface
+from .abstract_marker_interface import AbstractMarkerInterface
