+++ src/detector/structures/__init__.py
@@ -1,3 +1,4 @@
+from .calibration_configuration import CalibratorConfiguration
 from .calibration_image_metadata import CalibrationImageMetadata
 from .calibration_image_state import CalibrationImageState
 from .calibration_map import CalibrationMap
@@ -5,4 +6,6 @@ from .calibration_map_entry import CalibrationMapEntry
 from .calibration_map_value import CalibrationMapValue
 from .calibration_result_metadata import CalibrationResultMetadata
 from .calibration_result_state import CalibrationResultState
+from .camera_connection import CameraConnection
 from .charuco_board_specification import CharucoBoardSpecification
+from .detector_configuration import DetectorConfiguration
