+++ src/detector/structures/calibration_configuration.py
@@ -2,5 +2,4 @@ from pydantic import BaseModel, Field
 
 
 class CalibratorConfiguration(BaseModel):
-    serial_identifier: str = Field()
     data_path: str = Field()
