+++ src/detector/structures/calibration_image_metadata.py
