+++ src/detector/structures/calibration_image_state.py
