+++ src/detector/structures/calibration_map_entry.py
