+++ src/detector/structures/calibration_map_value.py
