+++ src/detector/structures/calibration_result_metadata.py
