+++ src/detector/structures/camera_connection.py
