+++ src/detector/structures/charuco_board_specification.py
